add flags field to MsgStream
Right now we only define the MSG_FLAG_MODE bit. This is used to
assert that a stream should attempt to force its write through in
the case of contention. If flag is set to 0 msg_write() will
return false when it detects contention has occurred. By setting
the MSG_MODE_MASTER flag when initializing the stream msg_write()
will attempt to resend the message header until it gets an
expected result (subject to the number of available retries).

An additional test was included for this case.