typedef the MsgStream function types
This makes it easier to change them in the future if needed (to
add custom allocator support for example).

The functions now take a pointer to the MsgStream itself instead
of the device member because it is too easy to foot gun yourself
and pass the wrong parameter when both the device and argument are
(void *). This is purely a programming aid since the compiler will
ignore all that and just pass an offset to the device member
anyways.